a first running version on Android, Crime-classes and db replaced by Entires and ViewModel classes, layout and layout classes remain to be replaced

+++ EntriesVM-lib/app/src/main/java/com/stefanbuehlmann/entriesvm/service/intf/EntryServiceI.java
@@ -8,7 +8,7 @@ import com.stefanbuehlmann.entriesvm.model.intf.EntryROI;
 import com.stefanbuehlmann.entriesvm.model.intf.EntryI;
 
 public interface EntryServiceI {
-    public long NO_ID = -1;
+    public int NO_ID = -1;
 
     public List<EntryI> selectWhere(String whereClause);
